# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CipNime Anime API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnimeItem {
  title: string;
  poster: string;
  episodes?: number | null;
  releaseDay?: string | null;
  latestReleaseDate?: string | null;
  animeId: string;
  score?: string | null;
  type?: string | null;
  status?: string | null;
}

export interface AnimeListResponse {
  animeList: AnimeItem[];
  hasNextPage: boolean;
  currentPage: number;
}

export interface GenreItem {
  title: string;
  genreId: string;
  href?: string;
}

export interface GenreListResponse {
  genreList: GenreItem[];
}

export interface EpisodeItem {
  title: string;
  eps: number;
  date?: string | null;
  episodeId: string;
  href?: string;
}

export interface AnimeDetailResponse {
  title: string;
  poster: string;
  japanese?: string | null;
  score?: string | null;
  type?: string | null;
  status?: string | null;
  episodes?: number | null;
  duration?: string | null;
  aired?: string | null;
  studios?: string | null;
  synopsis?: string | null;
  genreList: GenreItem[];
  episodeList: EpisodeItem[];
  batch?: string | null;
}

export interface StreamSource {
  server: string;
  url: string;
  quality?: string | null;
  isExternal?: boolean | null;
  type?: string | null;
}

export interface DownloadLink {
  server: string;
  url: string;
  quality?: string | null;
}

export interface EpisodeResponse {
  title: string;
  episode: string;
  sources: StreamSource[];
  downloadLinks?: DownloadLink[];
}

export type GetOngoingAnimeParams = {
  page?: number;
};

export type GetAnimeDetailParams = {
  animeId: string;
};

export type GetEpisodeSourcesParams = {
  episodeId: string;
};

export type SearchAnimeParams = {
  q: string;
  page?: number;
};

export type GetAnimeByGenreParams = {
  genreId: string;
  page?: number;
};