# lib/api-zod/src/generated/types/animeDetailResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CipNime Anime API
 * OpenAPI spec version: 0.1.0
 */
import type { EpisodeItem } from "./episodeItem";
import type { GenreItem } from "./genreItem";

export interface AnimeDetailResponse {
  title: string;
  poster: string;
  japanese?: string | null;
  score?: string | null;
  type?: string | null;
  status?: string | null;
  episodes?: number | null;
  duration?: string | null;
  aired?: string | null;
  studios?: string | null;
  synopsis?: string | null;
  genreList: GenreItem[];
  episodeList: EpisodeItem[];
  batch?: string | null;
}