# lib/api-zod/src/generated/types/animeItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CipNime Anime API
 * OpenAPI spec version: 0.1.0
 */

export interface AnimeItem {
  title: string;
  poster: string;
  episodes?: number | null;
  releaseDay?: string | null;
  latestReleaseDate?: string | null;
  animeId: string;
  score?: string | null;
  type?: string | null;
  status?: string | null;
}