# lib/api-zod/src/generated/types/animeListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CipNime Anime API
 * OpenAPI spec version: 0.1.0
 */
import type { AnimeItem } from "./animeItem";

export interface AnimeListResponse {
  animeList: AnimeItem[];
  hasNextPage: boolean;
  currentPage: number;
}