# lib/api-zod/src/generated/types/downloadLink.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CipNime Anime API
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadLink {
  server: string;
  url: string;
  quality?: string | null;
}