# lib/api-zod/src/generated/types/episodeItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CipNime Anime API
 * OpenAPI spec version: 0.1.0
 */

export interface EpisodeItem {
  title: string;
  eps: number;
  date?: string | null;
  episodeId: string;
  href?: string;
}