# lib/api-zod/src/generated/types/episodeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CipNime Anime API
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadLink } from "./downloadLink";
import type { StreamSource } from "./streamSource";

export interface EpisodeResponse {
  title: string;
  episode: string;
  sources: StreamSource[];
  downloadLinks?: DownloadLink[];
}