# lib/api-zod/src/generated/types/genreItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CipNime Anime API
 * OpenAPI spec version: 0.1.0
 */

export interface GenreItem {
  title: string;
  genreId: string;
  href?: string;
}