# lib/api-zod/src/generated/types/genreListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CipNime Anime API
 * OpenAPI spec version: 0.1.0
 */
import type { GenreItem } from "./genreItem";

export interface GenreListResponse {
  genreList: GenreItem[];
}